refactor sync client transport
