+++ tests/integration/dbapi/async/V1/test_queries_async.py
@@ -169,7 +169,6 @@ async def test_select(
         )
 
 
-@mark.skip(reason="V1 query doesn't finish within the timeout")
 @mark.slow
 @mark.timeout(timeout=600)
 async def test_long_query(
