+++ tests/integration/dbapi/sync/V1/test_queries.py
@@ -120,7 +120,6 @@ def test_select(
         )
 
 
-@mark.skip(reason="V1 query doesn't finish within the timeout")
 @mark.slow
 @mark.timeout(timeout=600)
 def test_long_query(
